% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_get_input.R
\name{aoc_get_input}
\alias{aoc_get_input}
\title{get AoC puzzle input}
\usage{
aoc_get_input(day, year = 2017, cookie_path, encoding = "UTF-8")
}
\arguments{
\item{day}{numeric. Day for which input is required}

\item{year}{numeric. Year for which input is required}

\item{cookie_path}{character string. Path to the cookie string stored as a text file.}

\item{encoding}{character string. Encoding of result. Defaults to "UTF-8".}
}
\value{
Returns user specific puzzle input as returned by the advent of code server.
}
\description{
Gets the user specific input for a given Advent of Code day puzzle directly
from the adventofcode.com server
}
\details{
In chrome, get your cookie details by clicking on the triple dot menu
(top right) > More tools > Developer tools > Application > Cookies. You need to copy the long
string in the \code{value} column for row \code{session}. Save that as a text file.
}
